refactor: Consolidate publisher functionality and remove obsolete modules; migrated EPUB conversion to bookworks/core/publisher and utility functions to bookworks/utils/helpers; updated app.py and tests accordingly; removed obsolete files: audio_gen.py, md_publish.py, clean_for_tts.py, bw_utils.py.
