--- bookworks/audio_gen.py
@@ -1,49 +1,0 @@
-"""
-Generate audio segments from markdown content using OpenAI Whisper.
-"""
-
-import os
-from typing import List, Tuple, Optional
-from dataclasses import dataclass
-
-
-@dataclass
-class AudioSegment:
-    """Represents a generated audio segment."""
-
-    file_path: str
-    duration: float
-    chapter_title: str
-    start_time: float
-
-
-class AudioGenerator:
-    """Handles the conversion of markdown content to audio segments."""
-
-    def __init__(self, output_dir: str = "audio_output"):
-        """
-        Initialize the audio generator.
-
-        Args:
-            output_dir: Directory where audio files will be saved
-        """
-        self.output_dir = output_dir
-        os.makedirs(output_dir, exist_ok=True)
-
-    def generate_audio(
-        self, content: str, voice: str = "default", chunk_size: int = 1000
-    ) -> Tuple[List[AudioSegment], Optional[str]]:
-        """
-        Convert markdown content to audio segments.
-
-        Args:
-            content: The markdown content to convert
-            voice: The voice style to use
-            chunk_size: Maximum number of characters per audio segment
-
-        Returns:
-            Tuple containing (list of AudioSegment objects, error_message)
-            If successful, error_message will be None
-        """
-        # TODO: Implement audio generation using OpenAI Whisper
-        raise NotImplementedError("Audio generation functionality is under development")
